only remove base.sdc if exists

+++ openlane/steps/__main__.py
@@ -426,7 +426,8 @@ def create_test(ctx, step_dir, output):
     step = load_step_from_inputs(ctx, None, config, state_in)
     step.create_reproducible(output, include_pdk=False, _flatten=True)
     os.remove(os.path.join(output, "run_ol.sh"))
-    os.remove(os.path.join(output, "base.sdc"))
+    if os.path.exists(os.path.join(output, "base.sdc")):
+        os.remove(os.path.join(output, "base.sdc"))
 
 
 @group(formatter_settings=formatter_settings)
